requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf_spectacular>=0.22.1,<0.23
cloudinary==1.38.0
django-cloudinary-storage>=0.3.0,<0.4
python-dotenv>=0.19.0,<0.20
Pillow>=9.0.0
django-cors-headers